More typing fixes
