+++ neurons/validator/rewards/models/__init__.py
@@ -1,3 +1,4 @@
+from neurons.validator.rewards.models.base import BaseRewardModel
 from neurons.validator.rewards.models.empty import EmptyScoreRewardModel
 from neurons.validator.rewards.models.blacklist import BlacklistFilter
 from neurons.validator.rewards.models.similarity import ModelSimilarityRewardModel
@@ -7,6 +8,7 @@ from neurons.validator.rewards.models.nsfw import NSFWRewardModel
 
 
 # Re-Export
+BaseRewardModel = BaseRewardModel
 EmptyScoreRewardModel = EmptyScoreRewardModel
 BlacklistFilter = BlacklistFilter
 ModelSimilarityRewardModel = ModelSimilarityRewardModel
